Add possibility to pass the session to last_path/3 and last_paths/2
Instead of just the conn, let these helpers also accept a map
representing the session. This because when in using Phoenix LiveViews
there is no conn available, but there is the session on mount. This
can be helpful whenever one would like to redirect to the last
non-LiveView path visited.